Cleaned master.cfg, factories.py. Deleted unused.
master.cfg is cleaned, there is still commented
examples of each step. factories.py is cleaned as
well, removed commented lines. Removed schedulers-
and builders.py, needed parameters for schedulers
and builders are stored in the projects.py and given
from there to the get_... functions, or directly
given when calling the get_... functions.