Major work on the labgrid/pytest
beagle1.yaml:
Changed prompt to be @arm instead of debian@
This is because part of the test commands must be
executed as root.

conftest.py:
This includes lines which allows to use --product
when running pytest command. This allows to supply
test scripts with a test target model and usage of
kernel_modules.py to loop through steps to install
modules, use overlay merger, and run tests.

helpers.py:
This file has a helper function that compares if
given string can be found in a given list.

test_init_overlay.py:
This is the first test to run after running the
test_shell.py (which tests if beagle bone booted
succesfully). test_init_overlay.py installs the
overlay merger to be used by device tree overlays
in the test_merge_dt_overlay.py

test_merge_dt_overlay.py:
This file merges given device tree binary file to
a device tree of the running kernel, and tests
that the operation was succesful.

test_insmod_tests.py:
This file is not functioning currently. Loops
for running the commands are in place, but this
is unfinished at the moment.

kernel_modules.py:
This is manually copied file from the buildbot
master. In time this will be transfered / updated
from master. This is the file that test____.py
files are using.