Changed commit message and tag checking
factories.py:

factories.py has helper functions isFloat(check)
and tagConvert(tagS) which together change the
release tag from string to a list of ints and
floats. These lists are easy to do logic and
maths with for comparing when a kernel version
should be tested on the test targets.
tagConvert(tagS) is called in the function
check_tag(step,target).

doStepIf needs a boolean or callable with 1
parameter, but two is needed.
check_tag(step,target) is a function wrapped with
check_tag_partial=functools.partial(check_tag,
target=target) to bypass the limitation of 1
parameter.

hideStepIf=skipped is added to all but first
skipped driver test step. skipped is a function
that returns True if results is 3. 3 is return
code for skipped.
-
master.cfg:

regex checking for commit message is changed to
looser version, which works for linux version 5.15
up to 99.99.999.