Changed kernel_modules dir, fixed func tagConvert
kernel_modules.py is moved to tests/driver_tests/configs
now. This change is to keep the driver_tests dir clean,
and to have all configs in same place as there will be
more and more config files for testing the targets.

function tagConvert(tagS) did not have functionality
to remove -rc from the tag. This caused version number
comparsion math to fail.