All allowed voltages for bd71837/47 are tested now
Added config file for bd71837 and test scripts for it.
-
pmic_conf.py:

def print_failures(self, failures):
This is a function that prints debug information to test the tests.

def regulator_is_on(self, regulator, command):
This function returns 1 if regulator is enabled, 0 if not.
For the special case in BD71847 and BD71837, this is done by checking
the device tree property with check_regulator_enable_mode function, and
is assumed to be on if regulator can not be turned off.

def regulator_voltage(self, regulator, r, command, volt_index=None):
Added if conditions to check if regulator has range selection. This
function has some repeating code which does not look nice, but it is
functional.

Now this function returns set microvoltage and calculated microvoltage
from received register value.

def regulator_voltage_run(self, regulator, command):
This function now returns a dictionary 'voltage_run', which contains
test_failed flag, and a list of failed regulator settings. These values
are set and appended if sent and received microvoltages differ.
-
test_voltage_run.py: (in bd71847/ and bd71837/)
Before running a regulator_voltage_run, it is checked if regulator
voltage can be changed.

Asserting is done with the flag 'test_failed' for testing purposes of
the tests, and if this flag is raised, buck fail info is also printed.
-
Modified test function names in test_*.py files to be the same as the
test file name.
-----------
Some cleaning up of white spaces and unneccessary commented lines.