Renamed tests. Added wled to bd71815, out_of_range
Renamed tests, so the executing order is alphabetical. This should be
needed when the tests gets added to the test automation process.

Added wled to bd71815.py:
This is as regulators, the test config file lists millivolts, but those
values are actually milliamperes.

Added test_003_out_of_range_voltages.py for each PMIC:
This test is functional, but for now the device trees are defined so
that the regulator framework correctly prohibits driver to even get the
information of a wronly requested voltage change.