bd71847 out of range test dts + modified Makefile
bd71847_test_outofrange.dts:
This file is a modified version of the bd71847_test.dts. This new file
has modifier regulator-min-microvolt and regulator-max-microvolt to
allow regulator framework to pass values that are out of allowed range
to the device driver.

Makefile for bd71847 modules:
Takes env variable {DTS_FILE} for choosing which .dts file is to be
used.