bd71815 .dts + Makefile for out of range test
bd71815_test_outofrange.dts:
This file is a modified version of the bd71815_test.dts. This new file
has modifier regulator-min-microvolt and regulator-max-microvolt to
allow regulator framework to pass values that are out of allowed range
to the device driver.

Makefile for bd71815 modules:
This file takes env variable {DTS_FILE} for choosing which .dts file is
to be used.