Refactoring pmic_conf.py
regulator_voltage_get(..) proved problematic. Refactoring this function
and functions related to it. New functions are called i2c_to_uv(..),
_i2c_to_volt_config(..)...

These changes are made to make the class easier to read and understand,
and to also have more reusable functions for calculating i2c values to a
form that the driver understands.