bd71847_test_oor_rampr1.dts added
This file has device tree properties so that regulator framework does
not block out of range (oor) settings so that driver can be tested to
work. Properties to check half of the ramp rate settings for BUCK1 and
BUCK2 are set as well.

Thought process here is that the default .dts file have descriptive name
for what they can test. Rest of the .dts files for further tests are
generated with a Python script and a config file.