Skip tests for config entries that have 'dts_only'
'ldodvref' and 'ldolprf' are features that regulators use, but in device
tree these are listed under regulators. Device tree generator is
checking config for every regulator found in the .dts, this caused the
device tree generator to fail.

Add exception to test scripts.

---
This might not be the most clever way to do this, but this works for
now.