pmic: generate_dts: Support to add properties
generate_dts() uses template dts file, which needs to have the
properties (as is, or commented out) in order to replace existing
property settings to generated dts file. This causes issues if
properties are added to config file which are not included in the
template file.

Add functionality to generate_dts() to insert properties at the end of a
regulator node.

Additional change: Use _print_property_to_dts() to check property type
and insert line to generated dts.