Tests: Report: Add functions to write results
Add functions to write results. Conditions to start writing results to
test report will be in the test stages them selves. helpers.py has
functions only for the generic stages that should be the same for other
products than the PMICs as well.

tf is most important parameter; it selects what to write to the test
report.

----
There is separate functions for writing results to file for the moment.
They will probably get cleaned off if none of those will be needed.