BuildBot: Add step to initialize test result
BuildBot: Add step to run_driver_tests() that uses initialize_results.py
to create a test result file.

initialize_results.py: This is a simple file that calls
initialize_report() to create or replace temp_results.txt. This script
takes one argument and it is the BuildBot projecti name that launched
this script.