BuildBot: projects: Describe dict with comments
This commit is for anyone besides me trying to figure out how to use
this test automation setup. This commented projects dictionary will be
used in the documentation that is not yet done.

Additional change:
Increase tree stable timer to give a bit more slack so no builds will
start too early.