Git bisect: Launch bisect
Initial steps to make it possible to launch git bisect is done now.

Triggerable BuildBot Scheduler is needed, and a step that triggers that
scheduler.

Step that triggers the git bisect scheduler also sets a property
'git_bisecting' : True. This property is needed to skip steps such as
regular git steps and test reporting.