Git bisect: Work towards bisect good / bad
Git bisect needs atleast one 'git bisect good' and 'git bisect bad'.

This commit contains work towards making it possible to sequence these
commands in the builders '/build' directory which contains the linux
git.

Functions git_bisect_good() and git_bisect_bad() are both at the end of
linux_driver_test().

Build steps in git_bisect_good() are executed when none of the driver
tests fail. It might not be the most clever thing to actually do this:
It is possible for the moment that driver tests are skipped because of
for example there is connection error to one of the beagle bones, and
tests for the PMICs connnected to that beagle bone could have failed but
the commit will get marked as good.

Build steps in git_bisect_bad() are executed when even one of the driver
tests across all PMICs fail.

Additional change here is that collecting dmesg and dts wont happen
twice if a step fails before launching 'dts' tests.

----
This is early stage commit towards making the automated git bisect work
well. Not a lot of the things are something that will remain the same
towards end. Don't use these changes as good examples for anything.