Git bisect: Move script for good commits to worker
There is not a way to set a property from MasterShellCommand on
BuildBot. Property is needed to pass the good commit to 'git bisect
good' command.

Move good_commits.py to worker directory and rename to
bisect_good_commit.py to give some sense on what that script is related
to.