Test results: Fix voltage run, change dir name
Voltage run results printed all voltage values for regulator that had
error. Only the failed ones are interesting. Wrong variable usage in if
statement causes this unintented behaviour.

Fix the if statement to use correct index that has the microvolt values
instead of everything in 'expect' and 'return'.

Additional change is to use 'temp_results' to write results while tests
are running. This is so that 'temp_results' can be moved to 'results'
at the end of the first build. If the build has an error, BuildBot will
trigger new builds with new checkout from git bisect, and that would
cause overwrites. Only the initial test reports are desired + output of
the gitbisect.