BuildBot: Add return code check to extract_fn
When steps fails because of connection issue, the 'FAILURES' print does
not work. In this case pytest exits with return code 1, so the return
code can be used to set properties.

Add if statement for return code different than 0 to set properties to
skip tests.

Additional change here is to use report_janitor.py to copy results. This
change makes the code more uniform.
----
Changes related to report_janitor.py are not tested yet.