Git bisect: Bisect if all beagles fail to login
It is possible that a kernel update breaks our test setup in a manner
that the kernel will not properly boot up on BeagleBone Black.
BeagleBone Black has a boot issue so it is possible that the login fails
for a single BeagleBone Black but that does not necessarily mean that
the kernel is broken ( eventhough 5 consecutive boot fails is highly
unlikely).

Add functionality to BuildBot to start git bisect if all BeagleBones
used for the test build fail to login