BuildBot: git bisect: Exclude linux-next
Git bisecting linux-next is problematic. It has non-linear history and a
tested change might break the test for a reason that is already fixed.

Exclude linux-next from triggering and running git bisect.

Later this can be added back with modifications, such as only marking
commit bad for the exact same issue that triggered the git bisecting.