Tests: generate test steps: Add exclusion for .swp
While developement of the system, it is often the case that one of the
test files is open with vim, which saves the buffer to .swp file. These
files cause generate_steps.py to generate BuildBot steps from those
files and ofc it will not work.

Add exclusion for .swp files in the generate_steps.py