BuildBot: Fix dtc compile command for BBB dtb
BuildBot uses projects' 'builderNames' to find the .dts file and specify directory
where to place the binary. This does not work when building the .dtb
with for Linux Stable versions because the projects' builder names are
appended with the linux stable branches.

Fix dtc command to use relative path instead.