Kernel config: arm32: Add lines for BD96801
arm32.config is generated with 'make menuconfig' from older kernel which
does not have BD96801.

Modules for BD96801 are manually
This is a quick and dirty fix to get a working kernel config that works
with Linux v6.11 and BD96801 as well.

----
It is planned that there will be better separate configs for different
kernel branches. For example Linux stable branches are vastly different
kernel versions, Linux Mainline and next can maybe use the same config
unless there is something very newly added to linux-next too (which
ofcourse will be the case).