BuildBot: Push results to git repository
factories.py:

Modify extract_get_timestamp():
Change result directory from results/ to test-results/

Add publish_results_git():
This is the function that adds BuildBot step to use report_janitor.py
which handles the commands to push results to git.

Add doStepIf_git_push():
This function returns True when the results should be published and
false when not. This is so that results are not published when git
bisect is triggered.

Add publish_results_git() to linux_driver_test():
linux_driver_test() is the function that calls all the other functions
that make up the factory to test drivers.

report_janitor.py:

Add argument 'publish_results_git':
When report_janitor.py is launched with 'publish_results_git', the
current builds result directory is pushed to git origin.

Arguments needed for this are:
	- date (must be the same property used when organizing results)
	- builder name (current projects builder name)
	- branch on origin where results should be pushed to

Full command looks like this:
'python3 report_janitor.py publish_results_git 2024_09_20_093414
Test_Linux test_push'

----
For the moment a temporary result report repository is in use.