BuildBot: Kunit: Separate kunit from driver tests
Kunit tests are done for each PMIC, but Kunit tests are not related to
the PMICs. Kunit tests are tests for the systems that drivers relies on.

Separate Kunit tests from driver tests, and do them independently if
kernel version has those included.

----
The current stage even with this commit has the following problem:
modprobe always returns 0 when Kunit module is installed. Success
assertion of the Kunit tests has to be done differently.