BuildBot: Clean up database: Add sqlite3 VACUUM;
"sqlite3 state.sqlite < cleanupdb.sql" clears info from the database,
but the size on the disk remains the same. This might lead to problems
eventually if the filesize keeps growing despite of clearing the entries
in the database. VACUUM rebuilds the database file repacking it into
minimal amount of disk space.

Add command to run sqlite3 VACUUM; after removing entries from the
database.