Tests: test_util: Sensor specific functionality
Test scripts use test_util to write and assert the test results.
Sensor tests are missing this functionality from the test_util.

Some test results are asserted against a range of values. This range is
specified with result['expect_high'] and result['expect_low']. If this
is the case, result['expect'] should be 'range'

The _assert_test() function now checks if the expected return value is a
value range. This function first closes the report files as always, and
after that checks if the result is a specific value or range of values
and does the assert call after that.

result['type'] can now be 'Sensor'.

For now sensors have assert/ report write functions for test_gsel and
sampling_frequency_match_timestamp tests. These are used by calling the
check_result() with a result dictionary where result['stage'] is set to
'test_gsel' or 'test_sampling_frequency_match_timestamp'.

Sensor tests are in early stage of development so this information might
change.