Tests: Sensor: KX022ACR-Z: Update sample rate test
Functions to do sampling rate tests have been changed in manner that it
can be defined in the test script how many samples are taken, and
tolerance is absolute percent instead of decimal value.

Change 'tolerance'  decimal value to percent value.

Use a function to increase the timestamp reading count value related to
the tested sampling frequency. This function uses math.ceil() to round
up. A normal round could be used, but this is the one that atleast never
rounds to 0.

Add function call to set watermark before reading timestamps.

Note that a dummy read must be done before setting watermark, as
described in the test scripts comment. Eventually this is not always
necessary to do the dummy_read, if there is tests that do reads before
setting the watermark. In this development phase it is time saving.

----
Maybe it is good idea to consider doing a dummy read in sanity check,
then test setting watermark and set it back to where it was.