Tests: test_util: Modify assert + samplerate funcs
Sampling rate test returns a list of values to compare to expected
value range. This can not be done by comparing the result['return']
to result['expect_high'] and result['expect_low'].

Add checks to _assert_test() to check if the result['return'] is a list,
and if so, assert that the retun value is in the expected range.

Sampling rate test returning a list causes a need to change to the
result writing function as well.

Modify result printing to use a loop structure to write results if a
return value in the result['return'] list is not in the allowed range.