BuildBot: Refactor 'projects' + creating factories
Factory instances (util.BuildFactory()) are initialized in factories.py,
which is kind of intuitive place for that. One technical problem this
way has is that the factory instance that a project uses is pointed in
the projects{}, in the projects.py thus the factories.py needs to be
imported to projects.py and projects.py needs to be imported in the
factories.py to create the build steps to that projects, factory
instances needs to be defined in factories.py before importing
projects.py and this is not good.

Refactoring linux stable branches to separate projects made it need
separate build factories as well and these factory instances are
missing.

Move creation of factory instances (util.BuildFactory()) to projects.py

Use for loop to create factories for linux stable

Fix linux_stable factory creation to happen in the same for loop which
creates the project dictionaries.

----
globals() is a python default function which retuns the dictionary of
global variables. These variables can be modified directly to the
dictionary. (As I understand, the global variables in python are stored
in this globals dictionary, variable names are keys, and values are
values.)

So line such as:
globals()['factory_linux_'+stable_branch = util.BuildFactory()
Has the same functionality as:
some_dictionary['factory_linux_'+stable_branch = util.BuildFactory()
Only the globals() is the method to access the dictionary containing the
global variables.