Tests: Sensor: KX022ACR-Z: Add G-scale raw match test
KX022ACR-Z has functionality to switch the acceleration range it
reports and driver supports it. This functionality needs to be tested.

kx022acr_z.py:
['axis']['bits'] is not very descriptive of what the value is.

Change ['bits'] to ['reg_bits'].

This value tells the register length for axis channels.

sensor_class.py:
To test this, test_gscale_raw() is called from the sensor class.
test_gscale_raw() takes following arguments:
  - axis		: (x, y, z) to be tested

  - test_type='match'	: Test type to be tested, rest of the function
			  can be used for other type of tests too.

  - tolerance=3		: Tolerance in percent.

  - append_results=False: When set to != False, more channel read
			  results can be appended to same result{}
			  so those results can be asserted in the same
			  test. Tolerance can be specified for each
			  channel separately.

test_gscale_raw() turns the raw data to signed, so that the values close
to 0 stay close to 0 if the sensor reads a negative number. Else the
reading in binary form jumps to something like 0b100000001010 which is
not close to 0 when asserting that the driver reading is close to
register read.

init_test_gscale() is currently used for developing tests. It has
potential to be used at the start of test_gscale_raw() and other similar
test functions. This function might get removed, but though here is to
make sensor_class.py flexible and possibly function for other type of
IIO devices than accelerometers. All in all, this function is currently
in use, but not in the tests that are ready + there is possible future
use.

driver_read_raw_xyz():
  - disables all channels and enables a single for the reading
  - returns int instead of stdout[0] which is string type

_reg_read_xyz_16bit():
  - returns unsigned int instead of signed int

reg_read_raw_xyz():
  - updated 'bits' to 'reg_bits' corresponding to change made in the
    sensor config file

test_util.py:
Add _assert functions for test result['stage'] 'gscale_raw_match'

These functions are used to write the results to report files and assert
the results.

test_002_gscale_raw_match.py():
This is the executed test file to check that raw values read from
drivers match the ones in the register, using different G scaling
settings.

This file demonstates the append_result and tolerance arguments of the
test_gscale_raw().