Git bisect: Exclude linux_stable from bisecting
There is a problem which is difficult to solve with Linux stable.
The problem has two origins:

1. It is difficult to trigger a scheduler with branch name inserted to
   end of a string.
2. There can be only one git poller per git repository.

First attempt was to put branch name to steps.Trigger schedulerNames.
This is where problem #1 was found.

Other tried solution was separate projects for each stable branch, but
that is where the problem #2 was bumped into.

-

Cleanest solution would be to figure out the problem #1. That should
require least amount of code change from 'master' branch.

Linux stable is a project with per branch schedulers and builders =>
multiple build directories. Build steps are generated per project basis.
So a factory of a single project should trigger a scheduler for the
same project, but use a builder for the branch that was build.