Kernel modules: Add test module for KX022ACR-Z
This is initial commit for a development branch focused on developing
test kernel module for Kionix KX022ACR-Z accelerometer.

kx022_acr_z_test.c is the source file for the kernel module. The kernel
module is utilizing IIO subsystem of the Linux kernel to read and write
information from/to the driver of KX022ACR-Z.

kx022a_i2c.dts has information to bind the test kernel module to driver
and access the accelerometers channels.

What works:
	- Read raw XYZ values
	- Read ms2 XYZ values
	- R/W buffer watermark
	- R/W sampling frequency
	- Read available sampling frequencies
	- R/W GSEL setting
	- Read available GSEL scales

What does not work / is not implemented:
	- Buffer reads
	- Reading timestamp channel

What is possible from the iio/consumer.h needs to be researched more.

The kernel module is currently being developed using Linux v6.12-rc5.

This commit includes bunch of shell scripts that helps with building and
deploying of binaries outside of the BuildBot driven test environment.