Test module: KX022A: Cleanup: Rename + add release
test_cleanup() is not very descriptive of what the function does, and it
is missing features to release the and unregister the registered buffer.

Change test_cleanup() to test_cleanup_buffer(). This naming is more
descriptive of what it does.

add iio_channel_release_all_cb() call to test_cleanup_buffer()