Test module: KX022ACR-Z: Attempt to read buffer
It is benefical to test reading the buffer. This is a bit messy state of
code, but some what functional.

Commented out struct scan from the cb function and moved to be global.

Added iio_channel_stop_all_cb() calls to _raw_show functions so that
the buffer do not have to be explicitly stopped.

_buffer_show() functions work, but they only print one value.

read_buffer_store() is an attempt to read and print multiple values from
buffer. buffer2list_store/show will be another attempt at this.