Test module: Accel: Generic accel module
The usage of test module made with KX022ACR-Z works with KX132ACR-LBZ.
Maybe at somepoint there will be the need to make changes to some
functions, or move the more generic functions to separate file.

Only changes to Makefile was needed. Script takes the sensor name in
'kx022acr_z' form to build the device tree and the test module.