Test module: Accel: Read 2 values from buffer
This is a first working version of managing to read anything from the
kfifo. I am sorry for the mess.

Callback function cb() copies data from the buffer and puts it in a
kfifo_accel with kfifo_in().

buffer2list reads data out from the kfifo_accel to output buffer which
is a type of struct scan as is the content of the buffer and
kfifo_accel.