Test module: Accel: Var names, _ATTR->DEVICE_ATTR
All structs and variables are matching generic_accel instead of
kx022acr_z. This change is to reflect that this kernel module is
functional for all accelerometer drivers currently in the test setup.

struct accel_priv has place for kfifo_accel pointer and watermark:
accel_priv is the struct type that is passed to functions to acces
private data.

struct accel_priv dev_priv is created in probe and passed to
dev_set_drvdata(), after which it is accessible with dev_get_drvdata()

Change from __ATTR to DEVICE_ATTR:
This change makes struct device_attributes, which' _show and _store
functions have struct device pointer as parameter. This allows to pass
private information to _show and _store functions.

Some of the unneeded and unused variables are now removed.