Test module: Accel: Use reinit_completion in cb()
It is unnecessary to call reinit_completion each time buffer is read and
if KFIFO has enough elements, complete() is called in the cb() right
after reading the buffer.

Remove reinit_completion() calls from the <xyz>_buffer_show() functions
and call it in the callback function cb() if the kfifo length is smaller
than where the watermark is set.