Helper scripts: iio_generic_buffer: cp src file
The helper script is compiles iio_generic_buffer that is in the linux
kernel source files. When getting new kernel source, the changes made to
iio_generic_buffer.c are lost and the script deploys an unmodified
version which does not work for automated testing.

Add copy command to script that replaces the iio_generic_buffer.c with a
modified version before building and deploying.