Test modules: Rename kx022acr-z_pd/ to generic_accel_test/
The kx022acr-z test module has become a generic module that works for
KX132ACR-LBZ and KX022ACR-Z. It is fitting to rename the directory at
this point, and it is somewhat necessary because starting to integrate
the accelerometer tests to the test automation system requires specific
directory paths.

This commit also fixes the paths in the helper scripts.

----
The deleted and new file changes are fixes to paths in the scripts.