BuildBot: accelerometer_factory: Copy test modules to nfs
Changes needed to make this happen are:

- Copy the original function with its doStepIf_ function to
  factory_helpers.py

- Modify the original copy_test_kernel_modules_to_nfs so that it takes
  factory as a parameter instead of project, and additional parameter is
  generic_testmodule which defaults to None.

Additional if, else control structure is added to do the old default
behaviour when the generic_testmodule is not given, else the generic
test kernel module is used that is given with this parameter.

	NOTE: This generic_testmodule specifies the directory from where
	      the module is copied from to nfs.

accelerometer_factory.py:
Call the copy_test_kernel_modules_to_nfs() from the factory_helpers.py
module with newly added parameters.
----

doStepIf_dts_report() location in the factory_helpers.py is moved to make
the code structure cleaner, now it is above the dts_report().