BuildBot: configs: test_boards.py: Re-organize structure
With split factories for accelerometers and PMICs it is essential
information on which power port the BeagleBones that have certain type
of products are connected to.

Change information structure to have the test target type at the top
level, followed by power port where the beagles are connected to.

Changed nformation structure make it possible to do following loop
structure:

for power_port in test_boards['accelerometer']['power_ports']
  for test_board in test_boards['accelerometer']['power_ports']
						    [power_port]

    for product in test_boards['accelerometer']['power_ports']
						    [power_port]
							[testboard]
	do_this()

product_types dictionary is not needed
Delete product_types dictionary