BuildBot: Make steps work with modified test_boards.py
test_boards.py informantion structure is changed and steps that use that
information do not work anymore.

Change steps to get information from the correct place.

In addition:
Do a sanity check before starting accelerometer factory.
This sanity check checks that iio_generic_buffer is in the NFS, as
without that the tests cannot be run.

Change test_000_check_iio_generic_buffer.py to run a proper command.
That '/joo' was used for debugging purpose.