BuildBot: Add db cleanup factory and scheduler
The deprecated factories.py has last step that cleans up the database.
And this feature is currently missing.

Add a nightly scheduler that runs once a week, that launches factory
which runs the same db cleaning command as before, on the BuildBot
master.

Priority is set to 1. Lower number is lower priority and higher number
is higher priority. This scheduler should have the lowest priority of
all.