Tests: Generic: Fix insmod_accel_test
test_003_insmod_accel_tests.py is a test scripts that installs the
kernel modules and checks that they were correctly installed.

To do this with 'modprobe' the kernel modules needs to be inside the
/lib/mobdules/<specific to kernel version>/kernel/

This script does not have a complete copy command to actually copy the
modules.

Fix copy command.

This script is missing the result_dir parameter, and this is one of the
generic test scripts that must use it.

Add result_dir parameter and insert it to result['result_dir'].