Tests: test results: report_janitor: separating result dirs
Add an option to report_janitor.py script that copies the results of a
Linux build for PMIC and accelerometer factory. This step is to make it
possible for concurrent runs of two factories to write results without
messing the summary.txt, and also pushing to separate git repositories
will be easier.

Utilize the <type> argument to initialize per product results when using
the script iwth 'initialize_product' argument. This change makes a
directory for that specific product inside the factory specific
directory.

Utilize the <type> argument to move and rename the temp results to
product specific directory and product specific file name.