BuildBot: Linux build factory: Fixes + sanitycheck()
Updated version of BuildBot does not include tags by default. This
causes problems in the test setup across different parts, such as
writing the results.

Add 'tags=True' to git step that updates the Linux kernel sources.

-

Add function that adds step to trigger PMIC test factory. It is not yet
timely to start testing the PMIC factory but soon enough it is, and this
is the first step towards including PMIC tests in the refactored
BuildBot.

-

Results are written to factory specific result directories, and they
do not have results of Linux build.

Add a function which adds a step that utilizes report_janitor.py to copy
the results of Linux build to factory specific result directories.

-

Copy the content of  kunit run_kunit_tests from deprecated factories.py,
and check_iio_generic_buffer and make sanitychecks():

sanitychecks() function logs into one BeagleBone Black from the
accelerometer test_boards and runs kunit tests and checks if the
iio_generic_buffer is present or not.