BuildBot: Change boolean properties to strings
Setting boolean or numeric properties for build triggered by
'steps.Trigger' with util.Property(<property>) sets the properties to
'null' value for whatever reason and that is not very helpful.

For unified style of properties and possibly less thinking in the
future - stringify all properties to have boolean True / False value to
'True' / 'False'.

----

Additionally:
doStepIf_initialize_product and doStepIf_dts_test_preparation have a new
if statement checking if factory is 'accelerometer' and if
iio_generic_buffer is found on the NFS.