BuildBot: Skip accelerometer if they can't be tested
Power down step for tests beagle bone and the step for finalizing
products' report file are still done even if they are not tested in the
first place.

Skip power down beagle bone and report finalizing step if
'iio_generic_buffer' is not found on the NFS, and do this conditionally
only when the factory doing these steps is 'accelerometer' factory.