BuildBot: Change workers to be LocalWorkers
LocalWorkers are a version of BuildBot workers that are present in the
master module, and they can be used when 'Master' and 'Worker' are on
the same machine.

Test setup runs in a single test pc, and to sequence different parts of
the test, multiple workers seems to be the way to go to have clear
limits how the builds can be started.

This change splits the worker1 to Test_Worker, and Linux_Worker:

Linux_Worker is used to build the linux and prepare everything for the
tests.

Test_Worker has the components used to execute the driver tests and
write results. This includes LabGrid configuration files and the test
scripts.