BuildBot: master.cfg: Add PMIC scheduler + clean up
Add get_pmic_scheduler() to add a scheduler in same manner as other
schedulers are added.

Add get_pmic_builder() to add a builder for in same manner as other
builders are added.

Some schedulers takes <beagle> parameter but it is not used at all.
These were added when parallel tests for each BeagleBone was though, but
it is not the process at the moment.

Remove beagle arguments and parameters for schedulers.