Tests: results: write temp results in /tmp
Temporary directories and files to write results are visible in the
filesystem with other Test_Worker related files and this creates
clutter.

Fix this by changing the filepaths where to create temporary directories
and files to /tmp/rohm_linux_driver_tests/ instead of path relative to
test setup internals.

It is not the most clever way to use this same path separately
everywhere, but for now this will do. When the
'/tmp/rohm_linux_driver_test' is written in paths.py and imported from
there, or even defined in multiple files separately it will reduce code
clutter too.