Tests: Sensor: Fix sampling frequency tests
There is some a notch in timestamps between last sample from reading
buffer to a watermark, and first sample of next set of samples. This
causes tests to fail.

Fix tests to use milliseconds as tolerance instead of percentual
fluctuation.

----

Old code is commented out, it might be useful so it is not removed
completely for now.