BuildBot: Fix Linux stable polling and testing
There has been various problems with Linux stable when it comes to
testing it. Such as only one gitpoller can be used per git repository.

When buildbot makes directories, it will replace the '.' with a '_', and
that causes problems.

Fix linux stable by adding a single project for it, but multiple
factories.

Fix linux stable worker directory accessing with BuildBot by replacing
the '.' in the branch name with '_'.

----
For now the git url is my personal repo where is couple of linux stable
branches for testing purposes.