Configs: kernel_modules: Fix KX022ACR-Z and add BD96801
KX022ACR-Z has wrong kernel version, this was changed to 6.12 when a bug
was found, but the fix is backported to Linux stable.

BD96801 configs are missing.

Fix correct kernel version for KX022ACR-Z.
Add BD96801.

----
BD96801 was in the test setup but these big refactoring changes + sensor
tests had 'wrong' branch as a base. At this point it is easier to just
re-add these lines.