BuildBot: linux factory: Fix execution of config steps
There is problems with using python if statements with refactored
factories, and the projects. Linux stable has a 'main project'
dictionary, which stores the branches and git repo address for polling.
Stable branches are build with own projects.

Change the kernel copy steps to use BuildBots doStepIf instead of python
if statement.

There is a bug in 'olddefconfig' step with the doStepIf value. It uses
Boolean false, when it should use stringified boolean 'True'.

Fix this by changing the value to match against.

Additionally:
	Add steps for: linux_mainline, linux_rohm_devel