Test results: Modify result dir FAILED/PASSED
Result directorys name is appended with _FAILED or _PASSED if anything
in the build goes wrong. This causes confusion if none of the PMIC tests
fail, but something fails in the Sensor tests.

Change it so that if Linux build / preparation step fails, mark
everything as failed. If only PMIC test fails or a test preparation step
in the PMIC factory, mark only PMIC results as failed. If Sensor test
fails, mark only Sensor results as failed.

This change required to split single_test_failed, to
pmic_single_test_failed and sensor_single_test_failed. To complete that
change, everything that used single_test_failed needed to be modified
accordingly.

Add new property: LINUX_RESULT
This property is used to mark everything as failed if something goes
wrong before launching the test factories.