Tests: Add tests for KX134-1211 accelerometer
This commit includes configuration <beagle6.yaml> for a 4-port FTDI chip
that is used to communicate with the beagle6, config for KX134-1211
including register information and samplerate / scale values, and tests
for the KX134-1211. Tests scripts are the same as for KX022ACR-LBZ with
a difference of scale values and some register addresses.