Test module: BD9576: Comment out hw_protection_shutdown
hw_protection_shutdown() is changed to hw_protection_trigger() in todays
linux-next. This causes compilation error with this file.

The hw_protection_shutdown / _trigger feature is currently not in use in
the test system, fastest way to make the test work with current and the
LTS Linux kernel versions is to comment out the bits where
hw_protection_shutdown() is used.

Comment out bits where hw_protection_shutdown() is used.