Results: Fix labeling results
If git bisect is started, the uploaded final results are labeled
depending on the success of the last build when they should be labeled
with the first build.

Make it so that the xxxx_RESULT properties are only set during the
initial build and carried over across the git bisect builds.