BuildBot: Add git bisecting for linux-next
linux-next git bisecting is a bit different than git bisecting a linear
branch. To get a good commit, the base of a tree must be first tested to
see if a good commit in that days git tree can be set. The base of the
tree is the head of linux-next/stable branch. After that the git
bisecting can really start in terms of this test system.

If a failure is detected in linux-next, 'git checkout stable' and launch
a "bisecting" build that either results in continued git bisecting, or
the bisecting ends because good commit is not found.

linux-next uses the git_bisect() function but it is branched for it
separately.
----
Additionally: fix 'git_bisecting' property from False to 'False' like
other boolean properties.